[RPC] Refactor all RPC code files in a seperate directory
