+++ src/dynamic-cli.cpp
@@ -11,8 +11,8 @@
 
 #include "chainparamsbase.h"
 #include "clientversion.h"
-#include "rpcclient.h"
-#include "rpcprotocol.h"
+#include "rpc/client.h"
+#include "rpc/protocol.h"
 #include "support/events.h"
 #include "util.h"
 #include "utilstrencodings.h"
