+++ src/httprpc.cpp
@@ -13,8 +13,8 @@
 #include "chainparams.h"
 #include "httpserver.h"
 #include "random.h"
-#include "rpcprotocol.h"
-#include "rpcserver.h"
+#include "rpc/protocol.h"
+#include "rpc/server.h"
 #include "sync.h"
 #include "ui_interface.h"
 #include "util.h"
