+++ src/init.cpp
@@ -55,8 +55,8 @@
 #include "consensus/validation.h"
 #include "privatesend-server.h"
 #include "psnotificationinterface.h"
-#include "rpcregister.h"
-#include "rpcserver.h"
+#include "rpc/register.h"
+#include "rpc/server.h"
 #include "scheduler.h"
 #include "script/sigcache.h"
 #include "script/standard.h"
