+++ src/qt/bdaplinkdetaildialog.cpp
@@ -6,9 +6,9 @@
 #include "ui_bdaplinkdetaildialog.h"
 
 #include "guiutil.h"
-#include "rpcregister.h"
-#include "rpcserver.h"
-#include "rpcclient.h"
+#include "rpc/register.h"
+#include "rpc/server.h"
+#include "rpc/client.h"
 #include "util.h"
 
 #include <stdio.h>
