+++ src/qt/dynamic.cpp
@@ -29,7 +29,7 @@
 
 #include "chainparams.h"
 #include "init.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "scheduler.h"
 #include "ui_interface.h"
 #include "util.h"
