+++ src/qt/test/rpcnestedtests.cpp
@@ -7,8 +7,8 @@
 #include "chainparams.h"
 #include "consensus/validation.h"
 #include "validation.h"
-#include "rpcregister.h"
-#include "rpcserver.h"
+#include "rpc/register.h"
+#include "rpc/server.h"
 #include "rpcconsole.h"
 #include "test/testutil.h"
 #include "univalue.h"
