+++ src/rest.cpp
@@ -10,7 +10,7 @@
 #include "httpserver.h"
 #include "primitives/block.h"
 #include "primitives/transaction.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "streams.h"
 #include "sync.h"
 #include "txmempool.h"
