+++ src/rpc/blockchain.cpp
@@ -16,7 +16,7 @@
 #include "instantsend.h"
 #include "policy/policy.h"
 #include "primitives/transaction.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "streams.h"
 #include "sync.h"
 #include "txdb.h"
