+++ src/rpc/dynode.cpp
@@ -17,7 +17,7 @@
 #include "privatesend-client.h"
 #endif // ENABLE_WALLET
 #include "privatesend-server.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "util.h"
 #include "utilmoneystr.h"
