+++ src/rpc/fluid.cpp
@@ -1,23 +1,23 @@
 // Copyright (c) 2017-2019 Duality Blockchain Solutions Developers
 
-#include "fluid.h"
 
 #include "bdap/domainentrydb.h"
 #include "bdap/utils.h"
 #include "chain.h"
 #include "core_io.h"
 #include "dynode-sync.h"
-#include "banaccount.h"
-#include "fluiddb.h"
-#include "fluiddynode.h"
-#include "fluidmining.h"
-#include "fluidmint.h"
-#include "fluidsovereign.h"
+#include "fluid/banaccount.h"
+#include "fluid/fluid.h"
+#include "fluid/fluiddb.h"
+#include "fluid/fluiddynode.h"
+#include "fluid/fluidmining.h"
+#include "fluid/fluidmint.h"
+#include "fluid/fluidsovereign.h"
 #include "init.h"
 #include "keepass.h"
 #include "net.h"
 #include "netbase.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "spork.h"
 #include "timedata.h"
 #include "util.h"
