+++ src/rpc/mining.cpp
@@ -21,7 +21,7 @@
 #include "miner/miner.h"
 #include "net.h"
 #include "pow.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "spork.h"
 #include "txmempool.h"
 #include "util.h"
