+++ src/rpc/register.h
