+++ src/rpc/server.h
@@ -9,7 +9,7 @@
 #define DYNAMIC_RPCSERVER_H
 
 #include "amount.h"
-#include "rpcprotocol.h"
+#include "rpc/protocol.h"
 #include "uint256.h"
 
 #include <univalue.h>
