+++ src/test/test_dynamic.cpp
@@ -16,8 +16,8 @@
 #include "net_processing.h"
 #include "pubkey.h"
 #include "random.h"
-#include "rpcserver.h"
-#include "rpcregister.h"
+#include "rpc/server.h"
+#include "rpc/register.h"
 #include "txdb.h"
 #include "txmempool.h"
 #include "ui_interface.h"
