+++ src/validation.cpp
@@ -40,7 +40,7 @@
 #include "primitives/block.h"
 #include "primitives/transaction.h"
 #include "random.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "script/script.h"
 #include "script/sigcache.h"
 #include "script/standard.h"
