+++ src/wallet/wallet.h
@@ -13,6 +13,7 @@
 #include "amount.h"
 #include "base58.h"
 #include "bdap/linkstorage.h"
+#include "rpc/wallet.h"
 #include "streams.h"
 #include "tinyformat.h"
 #include "ui_interface.h"
@@ -20,7 +21,6 @@
 #include "utilstrencodings.h"
 #include "validationinterface.h"
 #include "wallet/crypter.h"
-#include "wallet/rpcwallet.h"
 #include "wallet/wallet_ismine.h"
 #include "wallet/walletdb.h"
